feat / adding shopping cart item and shopping cart DTOs

+++ wigsshopping/src/main/java/com/naifer/wigsshop/wigsshopping/shoppingcartitems/ShoppingItemService.java
@@ -1,5 +1,6 @@
 package com.naifer.wigsshop.wigsshopping.shoppingcartitems;
 
+import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -8,6 +9,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import com.naifer.wigsshop.wigsshopping.productcategories.exception.ProductCategoryNotFoundException;
+import com.naifer.wigsshop.wigsshopping.productcategories.exception.ProductNotFoundException;
+import com.naifer.wigsshop.wigsshopping.utils.IImageService;
 
 @Component
 public class ShoppingItemService {
@@ -15,12 +18,32 @@ public class ShoppingItemService {
 	@Autowired
 	private ItemRepository itemRepository;
 	
+	@Autowired
+	private IImageService imageService;
+	
 	public List<ShoppingCartItem> getItems(){
 		List<ShoppingCartItem> items = itemRepository.findAll();
 		
 		if(items.isEmpty())
 			return List.of();
 		
+//		return items.stream()
+//					.map(item -> {
+//						ShoppingCartItemDTO itemDTO = new ShoppingCartItemDTO();
+//						
+//						itemDTO.setPrice(item.getProduct().getPrice());
+//						itemDTO.setId(item.getId());
+//						itemDTO.setQuantity(item.getQuantity());
+//						itemDTO.setTitle(item.getProduct().getTitle());
+//						itemDTO.setProductId(item.getProduct().getId());
+//						try {
+//							itemDTO.setProductImage(imageService.getImage(item.getProduct()));
+//						} catch (IOException e) {
+//							e.printStackTrace();
+//						}
+//						
+//						return itemDTO;
+//					}).toList();
 		return items;
 	}
 	
@@ -37,7 +60,14 @@ public class ShoppingItemService {
 		return itemRepository.save(item);
 	}
 	
-	public void deleteProduct(UUID itemId) {
+	public String deleteItem(UUID itemId) {
+		Optional<ShoppingCartItem> savedItem = itemRepository.findById(itemId);
+		
+		if(savedItem.isEmpty())
+			throw new ProductNotFoundException("id"+itemId);
+		
 		itemRepository.deleteById(itemId);
+		
+		return "Item successfully deleted.";
 	}
 }
